+++ lib/generators/rails_crud/templates/rails_crud.js
@@ -1,6 +1,6 @@
 function showCrudFileReloadMessage() {
     const message = document.createElement('div');
-    message.id = 'crud-file-reload-message';
+    message.id = 'rails-crud-reload-message';
     message.innerText = 'CRUD File Reloading...';
     message.style.position = 'fixed';
     message.style.top = '10px';
@@ -12,7 +12,7 @@ function showCrudFileReloadMessage() {
 }
 
 function hideCrudFileSyncMessage() {
-    const message = document.getElementById('crud-file-reload-message');
+    const message = document.getElementById('rails-crud-reload-message');
     if (message) {
         document.body.removeChild(message);
     }
@@ -22,7 +22,7 @@ function hideCrudFileSyncMessage() {
 // ジョブ実行中のメッセージを表示
 function showJobMessage() {
     const overlay = document.createElement('div');
-    overlay.id = 'job-message-overlay';
+    overlay.id = 'rails-crud-job-message-overlay';
     overlay.style.position = 'fixed';
     overlay.style.top = '0';
     overlay.style.left = '0';
@@ -35,7 +35,7 @@ function showJobMessage() {
     overlay.style.alignItems = 'center';
 
     const message = document.createElement('div');
-    message.id = 'job-message';
+    message.id = 'rails-crud-job-message';
     message.innerText = 'Job is running. Please do not operate the screen.';
     message.style.backgroundColor = 'yellow';
     message.style.padding = '20px';
@@ -48,7 +48,7 @@ function showJobMessage() {
 
 // ジョブ実行中のメッセージを非表示
 function hideJobMessage() {
-    const overlay = document.getElementById('job-message-overlay');
+    const overlay = document.getElementById('rails-crud-job-message-overlay');
     if (overlay) {
         document.body.removeChild(overlay);
     }
