+++ lib/generators/rails_crud/uninstall_generator.rb
